backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.31.0
python-dotenv==1.0.1
supabase==2.8.0
openai==1.109.1
python-pptx==1.0.2
matplotlib==3.9.2
seaborn==0.13.2
numpy==2.1.2
python-multipart==0.0.12
pydantic==2.9.2

#$env:Path += ";C:\Program Files\LibreOffice\program"